gps: fix size var

+++ modules/gps/gps.h
@@ -26,7 +26,7 @@ struct gps_msg {
 struct gps_handle_s {
     enum message_validity_t state;
     uint8_t parser_buffer[PARSER_BUFFER_SIZE];
-    uint8_t parser_cnt;
+    size_t parser_cnt;
 };
 
 bool gps_ubx_init_msg_topic(struct gps_handle_s* gps_handle, uint8_t class_id, uint8_t msg_id, void* frame_buffer, size_t frame_buffer_len, struct pubsub_topic_s* topic);
